Created separate file for each component
