--- web/cook-frontend/component_hierarchy.txt
@@ -1,15 +1,0 @@
-Describes the hierarchy of components
-
-     App
-      |
-   -------
-  |       |
-Header  Orders
-           |
-         ---------------
-         |       |     |
-        Order  Order  ...
-
-
-
-- 'Orders' holds state of all orders, each 'Order' will be passed it's content through a prop
