+++ web/cook-frontend/info/simple_react_style_guide.txt
@@ -1,0 +1,68 @@
+Many of the basic styling rules from Airbnb's style guide:
+Style guide: https://airbnb.io/javascript/react/
+Reference https://github.com/airbnb/react-dates for an example of how Airbnb implements their style guide
+
+
+One component per file
+- multiple stateless components are allowed in one file
+
+Extend React.Component, don't use React.createClass
+
+Components with no state should use normal functions, not classes or arrow functions
+
+.jsx extension for react components
+
+PascalCase for filenames
+
+PascalCase for React components and camelCase for their instances
+- import ReservationCard from './ReservationCard';
+- const reservationItem = <ReservationCard />;
+
+Use the filename as the component name
+
+Root component of a directory should be index.jsx
+
+Higher-order Component Naming <------- ?
+
+Don't use DOM component prop names for different purposes
+- Good: <MyComponent variant="fancy" />
+- Bad: <MyComponent className="fancy" />
+
+JSX:
+Multiple props both go on separate line
+- Good:
+<Foo
+  superLongParam="bar"
+  anotherSuperLongParam="baz"
+/>
+- Good: <Foo bar="bar" />
+
+Use double quotes for JSX attributes and single quotes for all other JS
+
+Include a single space in your self-closing tag
+- <Foo />
+
+Don't pad JSX attributes with spaces
+- Good: bar={baz}
+
+Use camelCase for prop names
+
+Avoid using array index as key prop
+- Good: key={item.id}
+- Bad: key={index}
+
+Define explicit defaultProps for all non-required props
+- Component.defaultProps = {bar: '', children: null};
+
+Use ref callbacks
+- Good: ref={(ref) => { this.myRef = ref; }
+- Bad: ref="myRef"
+
+Wrap JSX tags in parentheses when they span more than one line
+
+Self-close tags that have no children
+
+Bind event handlers for the render method in the constructor
+- don't use inline function
+
+Look at how to define propTypes, defaultProps, contextTypes. Search the document.
